+++ CHANGELOG.md
@@ -13,7 +13,7 @@
  - Can zoom to `nativeBbox` for `LayerentryComponent` and `LayerentryGroupComponent`.
 
 * **@dlr-eoc/map-ol:**
- - Try to udjust (native) bbox to enu before ol `setExtent`. This tries to fixes native bbox axis from wms 1.3.0. See `adjustBBoxAxisToEnu`.
+ - Transform extent in `setLayerExtentAfterProjection` if nativeBbox.epsg is not new epsg or try to use old extent.
 
 
 # [16.0.0](https://github.com/dlr-eoc/ukis-frontend-libraries/tree/v16.0.0) (2026-05-29) (update dependencies, new library ngx-ukis-ui-clarity and ngx-ukis-utilities, replace core-ui)